# docs/swagger.yaml
basePath: /
definitions:
  model.Category:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  model.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  model.CreateProductRequest:
    properties:
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    required:
    - name
    - price
    type: object
  model.Product:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
  model.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  model.UpdateProductRequest:
    properties:
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
info:
  contact: {}
  description: API untuk sistem kasir
  title: Go Kasir API
  version: "1.0"
paths:
  /categories:
    get:
      description: Ambil semua kategori
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Category'
            type: array
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Buat kategori baru
      parameters:
      - description: Category to create
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Category'
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Hapus kategori
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a category
      tags:
      - Categories
    get:
      description: Ambil kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
      summary: Get category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Perbarui kategori yang ada
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category data to update
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
      summary: Update an existing category
      tags:
      - Categories
  /health:
    get:
      description: Check API health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get health status
      tags:
      - Health
  /products:
    get:
      description: Ambil semua produk
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Product'
            type: array
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Buat produk baru
      parameters:
      - description: Product to create
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/model.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Product'
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      description: Hapus produk
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a product
      tags:
      - Products
    get:
      description: Ambil produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Perbarui produk yang ada
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
      summary: Update an existing product
      tags:
      - Products
swagger: "2.0"